Renames as format-campaigns
