+++ lib/middleware/campaigns-index/mobilecommons-groups.js
@@ -5,8 +5,8 @@ const logger = require('winston');
 const groups = require('../../groups');
 const helpers = require('../../helpers');
 
-function getGroupsForCampaign(campaign) {
-  return groups.getGroups(campaign.id, campaign.current_run)
+function getGroupsForPhoenixCampaign(campaign) {
+  return groups.getGroups(campaign.id, campaign.currentCampaignRun.id)
     .then((res) => {
       const groupsResponse = res.body.data;
       const scope = campaign;
@@ -25,7 +25,7 @@ function getGroupsForCampaign(campaign) {
 
 module.exports = function getMobileCommonsGroups() {
   return (req, res) => {
-    Promise.map(req.data, getGroupsForCampaign)
+    Promise.map(req.data, getGroupsForPhoenixCampaign)
       .then(data => res.send({ data }))
       .catch(err => helpers.sendErrorResponse(res, err));
   };
